chore: Reanmed ErrorData to ErrorType and moved to separate file
