+++ designsystem/src/main/java/au/com/alfie/ecomm/designsystem/component/dialog/error/ErrorType.kt
@@ -1,0 +1,7 @@
+package au.com.alfie.ecomm.designsystem.component.dialog.error
+
+data class ErrorType(
+    val message: String,
+    val buttonLabel: String,
+    var onButtonClick: (() -> Unit)? = null
+)
