+++ feature/shop/src/main/java/au/com/alfie/ecomm/feature/shop/ShopScreen.kt
@@ -24,7 +24,7 @@ import au.com.alfie.ecomm.core.navigation.arguments.wishlist.wishlistNavArgs
 import au.com.alfie.ecomm.core.ui.event.ClickEventOneArg
 import au.com.alfie.ecomm.core.ui.test.SERVICES_PAGE
 import au.com.alfie.ecomm.designsystem.component.bottombar.BottomBarState
-import au.com.alfie.ecomm.designsystem.component.dialog.ErrorData
+import au.com.alfie.ecomm.designsystem.component.dialog.error.ErrorType
 import au.com.alfie.ecomm.designsystem.component.segmented.SegmentedItem
 import au.com.alfie.ecomm.designsystem.component.segmented.SegmentedPage
 import au.com.alfie.ecomm.designsystem.component.snackbar.SnackbarCustomHostState
@@ -139,7 +139,7 @@ private fun ShopScreenContent(
                 onEvent = onWebViewEvent,
                 isBackHandlerEnabled = false,
                 modifier = Modifier.testTag(SERVICES_PAGE),
-                errorData = ErrorData(
+                errorType = ErrorType(
                     message = stringResource(R.string.error_failed_to_load_page),
                     buttonLabel = stringResource(R.string.retry)
                 )
